MEDIUM: rpmbuild: Do not use spectool for sources and patches download
The spectool version in el6 uses wget (not curl) which does not handle
the #/filename trick in URLs.

For instance the following RPM SPEC Source URL would be saved as
1.1.0.tar.gz with el6 spectool and rpmbuild would not find the file
because it would expect to find guacamole-server-1.1.0.tar.gz.

Source0: https://github.com/apache/guacamole-server/archive/1.1.0.tar.gz#/guacamole-server-1.1.0.tar.gz

Now that we have rpmspec available in our el6 image, we can use it
to extract the sources and patches URLs from the spec.