MAJOR: zsync: Set USE_RSYNC_UPDATE=1 in config to use rsync --update
This commit is tagged as MAJOR because it changes the default behavior
so that rsync is NOT called with --update, unless configuration variable
USE_RSYNC_UPDATE is set to 1.

By default, when we push / sync from the reference host, we expect
all files that are not in sync on remote hosts to be copied.