FEATURE: rpmbuild: Add el9 support
Base docker image rockylinux:9. On an old fc30 I have troubles with
seccomp filters due to the recent libc on el9. This is why it builds
with --security-opt seccomp=unconfined on docker run. Since that
option is not available in docker build, the rpmbuild/el9 image is
created manually with a script that terminates with docker container
commit.