BUG/MINOR: zpki: Fix ca-update-crt to ease regenerating a cert
If FORCE_CREATE_CRT is not set, certificate will be revoked but not
recreated because the file already exists. Since this action is ment
to recreate the certificate, FORCE_CREATE_CRT=1 shoule always be passed.

When we want to regenerate a certificate, for instance when SANs were
missing, SANs values passed on the CLI are ignored because a CSR
already exists and get reused. Passing CLI option -Fcsr will set
FORCE_CREATE_CRT=1, allowing to regenerate the CSR, hence creating a
new certificate that honors the SANs given on the CLI.